+++ src/main/web/js/fll-objects.js
@@ -21,9 +21,8 @@ function FinalistDBRow(categoryName, hour, minute, teamNumber) {
  * Used for packaging up and sending to the server to put in the database. Needs
  * to match fll.web.report.finalist.FinalistCategory.
  */
-function FinalistCategory(categoryName, isPublic, room) {
+function FinalistCategory(categoryName, room) {
   this.categoryName = categoryName;
-  this.isPublic = isPublic;
   this.room = room;
 }